# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CS546 API",
        "description": "Connecting clothing to a smart forum",
        "version": "0.0.1"
    },
    "host": "localhost:5000",
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "produces": [
        "application/json",
        "text/html"
    ],
    "paths": {
...
        "/forums/": {
            "get": {
                "summary": "Main Forum Page",
                "description": "View existing forums by most recent or most popular.",
                "tags": [
                    "Forum"
                ],
                "parameters": [
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "What order to view the forum posts in.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of forum posts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/forum"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create forum",
                "desciption": "Create new forum with desciption of forum and clothing tag",
                "tags": [
                    "Forum"
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "body",
                        "description": "Title of forum",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "clothing",
                        "in": "body",
                        "description": "Clothing tag of forum",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "body",
                        "description": "Id of user who is creating forum",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "desciption": "Successfully created a forum"
                    },
                    "301": {
                        "desciption": "Redirect to new forum page"
                    },
                    "400": {
                        "desciption": "Bad request"
                    },
                    "500": {
                        "desciption": "Internal server error"
                    }
                }
            }
        },
        "/forums/:forum_id": {
            "get": {
                "summary": "View post by id",
                "description": "View specific forum post by id along with all comments associated with it.",
                "parameters": [
                    {
                        "name": "forum_id",
                        "in": "parameter",
                        "description": "id of forum to view",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Forum"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully view forum post.",
                        "schema": {
                            "type": "text/html"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update forum",
                "description": "Update specific fields of forum",
                "tags": [
                    "Forum"
                ],
                "parameters": [
                    {
                        "name": "forum_id",
                        "in": "parameter",
                        "description": "id of forum to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "remove_clothing_tag",
                        "in": "header",
                        "description": "array of clothing tag to remove",
                        "required": false,
                        "type": "array",
                        "item": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "add_clothing_tag",
                        "in": "header",
                        "description": "array of clothing tag to add",
                        "required": false,
                        "type": "array",
                        "item": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated forum post.",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "400": {
                        "desciption": "Bad request",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "500": {
                        "desciption": "Internal server error",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete forum",
                "description": "Remove all forums and comments linked to that forum",
                "parameters": [
                    {
                        "name": "forum_id",
                        "in": "parameter",
                        "description": "id of forum to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Forum"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted forum post.",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "400": {
                        "desciption": "Bad request",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "500": {
                        "desciption": "Internal server error",
                        "schema": {
                            "type": "application/json"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/forums/:forum_id/comments/": {
            "get": {
                "summary": "Get posted comments.",
                "description": "Get a list of comments for that forum id.",
                "parameters": [
                    {
                        "name": "forum_id",
                        "in": "parameter",
                        "description": "id of forum to view",
                        "required": true,
                        "type": "string"
                    }
                ],
                "response": {
                    "200": {
                        "description": "Successfully posted a comment"
                    },
                    "400": {
                        "desciption": "Bad request"
                    },
                    "500": {
                        "desciption": "Internal server error"
                    }
                }
            },
            "post": {
                "summary": "Add comment to post.",
                "description": "Get a list of comments for that forum id.",
                "parameters": []
            }
        },
        "/forums/:forum_id/comments/:comment_id/": {
            "get": {
                "summary": "Get comment by id for specific post",
                "description": "Return the specified comment",
                "response": {
                    "200": {
                        "description": "Comment contents",
                        "schema": {
                            "type": "array",
                            "items": {
                                "user_id": "id of user who posted",
                                "forum_id": "id of original forum post",
                                "comment": "contents of comment",
                                "likes": "number of likes on comment",
                                "likes_users": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "desciption": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "summary": "Update a comment.",
                "description": "Get a list of comments for that forum id.",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "parameter",
                        "description": "id of forum to view",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
            
        },
        "/forums/:clothing_type/": {
            "get": {
                "summary": "Sub-forum for clothing type",
                "description": "Shows all forum posts under specified clothing type",
                "parameters": [
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "What order to view the forum posts in.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Forum"
                ],
                "responses": {
                    "200": {
                        "description": "An array of forum posts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/forum"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create forum",
                "desciption": "Create new forum with desciption of forum and clothing tag",
                "parameters": []
            }
        },
...
    },
    "definitions": {
        "forum": {
            "type": "object",
            "properties": {
                "forum_id": {
                    "type": "string",
                    "desciption": "Unique forum identifier"
                },
                "createdBy": {
                    "type": "string",
                    "desciption": "ID of user"
                },
                "createdOn": {
                    "type": "date",
                    "desciption": "date post was created"
                },
                "label": {
                    "type": "array",
                    "description": "Array of labels for forum"
                },
                "title": {
                    "type": "string",
                    "description": "title of post"
                },
                "content": {
                    "type": "string",
                    "desciption": "Content of post"
                },
                "clothing": {
                    "type": "array",
                    "item": {
                        "type": "string",
                        "definition": "clothing id array"
                    }
                },
                "comments": {
                    "type": "array",
                    "item": {
                        "schema": "#/definitions/comment"
                    }
                }
            }
        },
        "comment": {
            "type": "object",
            "schema": {
                "_id": "string",
                "datePosted": "date",
                "displayName": "string",
                "numberLikes": "int32"
            }
        },
...
        "error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}